moving BQ seed code directly into db/seeds.rb & tests
