+++ db/seeds.rb
@@ -180,3 +180,20 @@ BrandingGroup.create!(name: 'Test Brand', user_id: api_user.id, font_family: 'He
 PresetSearch.create!(name: 'Test Search', search_terms: ["Testing Study"],
                      facet_filters: ['species:NCBITaxon_9606', 'disease:MONDO_0000001'], accession_whitelist: %w(SCP1))
 FeatureFlag.create!(name: 'faceted_search')
+
+# seed BQ for search controller test
+puts "Directly seeding BigQuery w/ synthetic data"
+File.open(Rails.root.join('db', 'seed', 'bq_seeds.json')) do |bq_seeds|
+  bq_data = JSON.parse bq_seeds.read
+  bq_data.each do |entry|
+    entry['CellID'] = SecureRandom.uuid
+    entry['study_accession'] = study.accession
+    entry['file_id'] = metadata_file.id.to_s
+  end
+  Tempfile.open(['tmp_bq_seeds', '.json']) do |tmp_file|
+    tmp_file.write bq_data.map(&:to_json).join("\n")
+    table = ApplicationController.big_query_client.dataset(CellMetadatum::BIGQUERY_DATASET).table(CellMetadatum::BIGQUERY_TABLE)
+    table.load tmp_file, write: 'append'
+  end
+end
+puts "BigQuery seeding completed"
